Layout 2 labels side by side.
This commit is a big one. I decided that laying out a single lable in
portrait on a single Zink sheet didn't give enough benefit to the album
artwork size to be worth the wastage.

This commit rotates the canvas from portrait to landscape and shortens
the labels by a couple of mm to compensate.

It also allows 2 layouts to be arranged on the same label, and to avoid
the cluttering of cli options for this we can use a layout described in
toml.

see the example toml file in `res/config.toml` for an example